requirements.txt
pillow>=9.1.0